11° versione 4

+++ Archivio.php
@@ -5,34 +5,56 @@ class Archivio
 {
     protected $driver ;
 
-    public function __construct($driver)
+    public function __construct(DriverArchivio $driver)
     {
-        $this->driver = $driver;
+        // TODO: controllare $driver--> nome classe!!!!
+
+            $this->driver = $driver;
+
 
     }
 
 
-    public  function add($nome_file,$contenuto){
+    public  function add($nome_file,$contenuto)
+    {
+        // TODO: controllare che esista, o meglio che NON esista.
+        if (!$this->exists($nome_file)) {
+            return $this->driver->insert($nome_file, $contenuto);
+        }
+    }
+    public function del($nome_file)
+    {
+        // TODO controllare che ESISTA
+        if ($this->exists($nome_file)) {
+            $this->driver->delete($nome_file);
+        }
+    }
+    public function edit($nome_file,$new_contenuto)
+    {
+        // TODO controllare che esista
+        if ($this->exists($nome_file)) {
+            return $this->driver->update($nome_file, $new_contenuto);
+        }
+    }
+    public function get ($key)
+    {
+        $file = $this->driver->get($key);
 
-       return $this->driver->insert($nome_file,$contenuto);
+        if ($file) {
+
+            return $file;
+        }
+
+        return false;
+    }
+
+    public  function lista ()
+    {
+        return $this->driver->lista();
 
     }
-    public function del($nome_file){
-        $this->driver->delete($nome_file);
-
-    }
-    public function edit($nome_file,$new_contenuto){
-       return $this->driver->update($nome_file,$new_contenuto);
-
-    }
-    public function get($key){
-       return $this->driver->get($key);
-
-    }
-    public  function lista(){
-
-     return $this->driver->lista();
-
+    public function exists ($file){
+        return $this->driver->exists($file);
     }
 
 }
